requirements.txt
h5py~=3.8.0
numpy~=1.23.4
torch
opencv-python~=4.7.0.72
torchvision
scipy
pillow~=9.3.0
tqdm~=4.64.1
matplotlib~=3.5.1
pyyaml
pandas
jenkspy
scikit-learn
gdown
kaggle
